Refactor: clean du code
