+++ src/main/java/friutrodez/backendtourneecommercial/repository/mysql/UtilisateurRepository.java
@@ -2,7 +2,6 @@ package friutrodez.backendtourneecommercial.repository.mysql;
 
 import friutrodez.backendtourneecommercial.model.Utilisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
