+++ src/test/java/friutrodez/backendtourneecommercial/ConnexionTest.java
@@ -2,13 +2,10 @@ package friutrodez.backendtourneecommercial;
 
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
-import com.mongodb.client.MongoDatabase;
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.data.mongodb.core.MongoTemplate;
-
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
@@ -17,10 +14,15 @@ import java.sql.SQLException;
 public class ConnexionTest {
     @Value("${spring.data.mongodb.uri}")
     private String mongoUri;
+    @Value("${spring.datasource.url}")
+    private String mySQLURL;
+    @Value("${spring.datasource.username}")
+    private String userMySQL;
+    @Value("${spring.datasource.password}")
+    private String passwordMySQL;
     @Test
     void ConnexionMySQLTest() throws SQLException {
-
-        Connection conn = DriverManager.getConnection("jdbc:mysql://127.0.0.1:3306/mydatabase", "myuser", "secret");
+        Connection conn = DriverManager.getConnection(mySQLURL, userMySQL, passwordMySQL);
         Assert.assertNotNull(conn);
     }
